Updated check_ge so that it looks for other matching GE physiological files and adds the .gep extension to them too

+++ phys2bids/utils.py
@@ -123,17 +123,51 @@ def check_ge(filename, indir):
         or a fullpath to such folder
 
     """
+    from glob import glob
 
-    ge_types = ['PPGData', 'RESPData', 'ECGData']
+    ge_types = ['ECGData', 'PPGData', 'RESPData']
 
+    # Ensure the file exists
+    if not os.path.isfile(os.path.join(indir, filename)) and filename is not None:
+        raise FileNotFoundError(f'The file {filename} does not exist!')
+
+    # Check if it's a GE file and add file extension
+    # Do the same for other linked files
     if any(ge_type in filename for ge_type in ge_types):
+        LGR.info('GE physiological data detected.')
+        # Look for related GE files based on timestamp in name
+        fnames = glob(os.path.join(indir, f'*{filename[-20:]}*'))
+        # Add extension to original so it's logged appropriately
         if 'gep' not in filename[:-3]:
             new_filename = filename + '.gep'
             copy_file(os.path.join(indir, filename),
                       os.path.join(indir, new_filename))
-            LGR.info('GE physiological data detected. Appending ".gep" to filename.')
+            LGR.info(f'Appending ".gep" extension to {filename}.')
         else:
-            LGR.info('GE physiological data detected. ".gep" extension present. ')
+            LGR.info(f'".gep" extension already present in {filename}.')
+        # Add extension to additional files and log these
+        # Remove the original filename from the list
+        fnames.remove(os.path.join(indir, filename))
+        try:
+            fnames.remove(os.path.join(indir, filename+'.gep'))
+        except ValueError:
+            pass
+        # Log if there are no additional files
+        if len(fnames) == 0:
+            LGR.info('No additional GE physiological files found')
+        else:
+            LGR.info('Additional GE physiological file(s) found')
+            for fname in fnames:
+                print(f'GE physiological file {fname.split("/")[-1]} also found.')
+                if 'gep' not in fname[:-3]:
+                    new_fname = fname + '.gep'
+                    copy_file(os.path.join(indir, fname),
+                              os.path.join(indir, new_fname))
+                    LGR.info(f'Appending ".gep" extension to {fname.split("/")[-1]}.')
+                else:
+                    LGR.info(f'".gep" extension already present in {fname.split("/")[-1]}.')
+
+
 
 
 def copy_file(oldpath, newpath, ext=''):
